zephyr: Enforce uniqueness on device tree wrappers
Move all of the device implementations into a `zephyr::device` module.

In this module, add a `Unique` type that supports the constructors on
the device requiring a unique instance.

The device tree augmentation code adds a declaration for these
uniqueness markers for each instance, passing it into the constructor.
