+++ zephyr/src/sys.rs
@@ -78,133 +78,3 @@ pub mod critical {
         }
     }
 }
-
-pub mod gpio {
-    //! Most devices in Zephyr operate on a `struct device`.  This provides untyped access to
-    //! devices.  We want to have stronger typing in the Zephyr interfaces, so most of these types
-    //! will be wrapped in another structure.  This wraps a Gpio device, and provides methods to
-    //! most of the operations on gpios.
-
-    use crate::raw;
-
-    /// A single instance of a zephyr device to manage a gpio controller.  A gpio controller
-    /// represents a set of gpio pins, that are generally operated on by the same hardware block.
-    pub struct Gpio {
-        /// The underlying device itself.
-        #[allow(dead_code)]
-        pub(crate) device: *const raw::device,
-    }
-
-    impl Gpio {
-        /// Constructor, used by the devicetree generated code.
-        ///
-        /// TODO: Guarantee single instancing.
-        pub unsafe fn new(device: *const raw::device) -> Gpio {
-            Gpio { device }
-        }
-
-        /// Verify that the device is ready for use.  At a minimum, this means the device has been
-        /// successfully initialized.
-        pub fn is_ready(&self) -> bool {
-            unsafe {
-                raw::device_is_ready(self.device)
-            }
-        }
-    }
-
-    /// A GpioPin represents a single pin on a gpio device.  This is a lightweight wrapper around
-    /// the Zephyr `gpio_dt_spec` structure.
-    #[allow(dead_code)]
-    pub struct GpioPin {
-        pub(crate) pin: raw::gpio_dt_spec,
-    }
-
-    impl GpioPin {
-        /// Constructor, used by the devicetree generated code.
-        ///
-        /// TODO: Guarantee single instancing.
-        pub unsafe fn new(device: *const raw::device, pin: u32, dt_flags: u32) -> GpioPin {
-            GpioPin {
-                pin: raw::gpio_dt_spec {
-                    port: device,
-                    pin: pin as raw::gpio_pin_t,
-                    dt_flags: dt_flags as raw::gpio_dt_flags_t,
-                }
-            }
-        }
-
-        /// Verify that the device is ready for use.  At a minimum, this means the device has been
-        /// successfully initialized.
-        pub fn is_ready(&self) -> bool {
-            self.get_gpio().is_ready()
-        }
-
-        /// Get the underlying Gpio device.
-        pub fn get_gpio(&self) -> Gpio {
-            Gpio {
-                device: self.pin.port,
-            }
-        }
-
-        /// Configure a single pin.
-        pub fn configure(&mut self, extra_flags: raw::gpio_flags_t) {
-            // TODO: Error?
-            unsafe {
-                raw::gpio_pin_configure(self.pin.port,
-                    self.pin.pin,
-                    self.pin.dt_flags as raw::gpio_flags_t | extra_flags);
-            }
-        }
-
-        /// Toggle pin level.
-        pub fn toggle_pin(&mut self) {
-            // TODO: Error?
-            unsafe {
-                raw::gpio_pin_toggle_dt(&self.pin);
-            }
-        }
-    }
-}
-
-pub mod flash {
-    //! Device wrappers for flash controllers, and flash partitions.
-
-    use crate::raw;
-
-    #[allow(dead_code)]
-    pub struct FlashController {
-        pub(crate) device: *const raw::device,
-    }
-
-    impl FlashController {
-        /// Constructor, intended to be called by devicetree generated code.
-        ///
-        /// TODO: Instance safety
-        pub unsafe fn new(device: *const raw::device) -> FlashController {
-            FlashController { device }
-        }
-    }
-
-    /// A wrapper for flash partitions.  There is no Zephyr struct that corresponds with this
-    /// information, which is typically used in a more direct underlying manner.
-    #[allow(dead_code)]
-    pub struct FlashPartition {
-        /// The underlying controller.
-        #[allow(dead_code)]
-        pub(crate) controller: FlashController,
-        #[allow(dead_code)]
-        pub(crate) offset: u32,
-        #[allow(dead_code)]
-        pub(crate) size: u32,
-    }
-
-    impl FlashPartition {
-        /// Constructor, intended to be called by devicetree generated code.
-        pub unsafe fn new(device: *const raw::device, offset: u32, size: u32) -> FlashPartition {
-            // The `get_instance` on the flash controller would try to guarantee a unique instance,
-            // but in this case, we need one for each device, so just construct it here.
-            let controller = FlashController::new(device);
-            FlashPartition { controller, offset, size }
-        }
-    }
-}
